+++ tests/Fixture/YoutubeVideosFixture.php
@@ -19,7 +19,6 @@
  * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
- * @see         MeCms\Utility\SitemapBuilder
  */
 namespace MeCmsYoutube\Test\Fixture;
 
@@ -78,8 +77,8 @@ class YoutubeVideosFixture extends TestFixture
             'is_spot' => 0,
             'seconds' => 3939,
             'duration' => '01:05:39',
-            'created' => '2016-12-31 15:05:40',
-            'modified' => '2016-12-31 15:05:40',
+            'created' => '2016-11-30 15:05:40',
+            'modified' => '2016-11-30 15:05:40',
         ],
         [
             'id' => 2,