+++ tests/TestCase/Utility/SitemapTest.php
@@ -19,7 +19,6 @@
  * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
- * @see         MeCms\Utility\SitemapBuilder
  */
 namespace MeCmsYoutube\Test\TestCase\Utility;
 
@@ -97,7 +96,7 @@ class SitemapTest extends TestCase
             ],
             [
                 'loc' => 'http://localhost/video/1',
-                'lastmod' => '2016-12-31T15:05:40+00:00',
+                'lastmod' => '2016-11-30T15:05:40+00:00',
                 'priority' => '0.5',
             ],
         ];